Cleaned up navigation drawer
