+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsFragment.kt
@@ -37,6 +37,7 @@ open class SettingsFragment : XpPreferenceFragment() {
         addPreferencesFromResource(R.xml.pref_downloads)
         addPreferencesFromResource(R.xml.pref_sources)
         addPreferencesFromResource(R.xml.pref_sync)
+        addPreferencesFromResource(R.xml.pref_backup)
         addPreferencesFromResource(R.xml.pref_advanced)
         addPreferencesFromResource(R.xml.pref_about)
 
@@ -80,6 +81,7 @@ open class SettingsFragment : XpPreferenceFragment() {
             "downloads_screen" to R.drawable.ic_file_download_black_24dp,
             "sources_screen" to R.drawable.ic_language_black_24dp,
             "sync_screen" to R.drawable.ic_sync_black_24dp,
+            "backup_screen" to R.drawable.ic_backup_black_24dp,
             "advanced_screen" to R.drawable.ic_code_black_24dp,
             "about_screen" to R.drawable.ic_help_black_24dp
     )
