+++ app/src/main/res/menu/menu_navigation.xml
@@ -1,43 +1,45 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <group android:id="@+id/group_feature"
-           android:checkableBehavior="single">
-        <item
-            android:checked="true"
-            android:id="@+id/nav_drawer_library"
-            android:icon="@drawable/ic_book_black_24dp"
-            android:title="@string/label_library" />
-        <item
-            android:id="@+id/nav_drawer_recent_updates"
-            android:icon="@drawable/ic_update_black_24dp"
-            android:title="@string/label_recent_updates" />
-        <item
-            android:id="@+id/nav_drawer_recently_read"
-            android:icon="@drawable/ic_glasses_black_24dp"
-            android:title="@string/label_recent_manga"/>
-        <item
-            android:id="@+id/nav_drawer_catalogues"
-            android:icon="@drawable/ic_explore_black_24dp"
-            android:title="@string/label_catalogues" />
-        <item
-            android:id="@+id/nav_drawer_latest_updates"
-            android:icon="@drawable/ic_watch_later_black_24dp"
-            android:title="@string/label_latest_updates" />
-        <item
-            android:id="@+id/nav_drawer_downloads"
-            android:icon="@drawable/ic_file_download_black_24dp"
-            android:title="@string/label_download_queue" />
-    </group>
-    <group android:id="@+id/group_settings"
-           android:checkableBehavior="single">
+    <item android:title="@string/group_library">
+        <menu>
+            <item
+                android:id="@+id/nav_drawer_library"
+                android:checked="true"
+                android:icon="@drawable/ic_book_black_24dp"
+                android:title="@string/label_browse" />
+            <item
+                android:id="@+id/nav_drawer_recent_updates"
+                android:icon="@drawable/ic_update_black_24dp"
+                android:title="@string/label_recent_updates" />
+            <item
+                android:id="@+id/nav_drawer_recently_read"
+                android:icon="@drawable/ic_glasses_black_24dp"
+                android:title="@string/label_recent_manga" />
+        </menu>
+    </item>
+    <item android:title="@string/group_catalogue">
+        <menu>
+            <item
+                android:id="@+id/nav_drawer_catalogues"
+                android:icon="@drawable/ic_explore_black_24dp"
+                android:title="@string/label_discover" />
+            <item
+                android:id="@+id/nav_drawer_latest_updates"
+                android:icon="@drawable/ic_watch_later_black_24dp"
+                android:title="@string/label_latest_updates" />
+            <item
+                android:id="@+id/nav_drawer_downloads"
+                android:icon="@drawable/ic_file_download_black_24dp"
+                android:title="@string/label_download_queue" />
+        </menu>
+    </item>
+    <group
+        android:id="@+id/group_settings"
+        android:checkableBehavior="single">
         <item
             android:id="@+id/nav_drawer_settings"
+            android:checkable="false"
             android:icon="@drawable/ic_settings_black_24dp"
-            android:title="@string/label_settings"
-            android:checkable="false" />
-        <item
-            android:id="@+id/nav_drawer_backup"
-            android:icon="@drawable/ic_backup_black_24dp"
-            android:title="@string/label_backup" />
+            android:title="@string/label_settings" />
     </group>
 </menu>
