+++ app/src/main/res/values/strings.xml
@@ -8,12 +8,15 @@
     <string name="label_download_queue">Download queue</string>
     <string name="label_library">My library</string>
     <string name="label_recent_manga">Recently read</string>
-    <string name="label_catalogues">Catalogues</string>
-    <string name="label_recent_updates">Library updates</string>
+    <string name="label_browse">Browse</string>
+    <string name="label_discover">Discover</string>
+    <string name="label_recent_updates">Recent updates</string>
     <string name="label_latest_updates">Latest updates</string>
     <string name="label_categories">Categories</string>
     <string name="label_selected">Selected: %1$d</string>
     <string name="label_backup">Backup</string>
+    <string name="group_library">Library</string>
+    <string name="group_catalogue">Catalogue</string>
 
     <!-- Actions -->
     <string name="action_settings">Settings</string>
@@ -162,7 +165,13 @@
     <string name="fourth_to_last">Fourth to last chapter</string>
     <string name="fifth_to_last">Fifth to last chapter</string>
 
-      <!-- Sync section -->
+    <!-- Backup section -->
+    <string name="pref_create_backup">Create backup</string>
+    <string name="pref_create_backup_summ">Can be used to restore current library</string>
+    <string name="pref_restore_backup">Restore backup</string>
+    <string name="pref_restore_backup_summ">Restore library from backup file</string>
+
+    <!-- Sync section -->
     <string name="services">Services</string>
 
       <!-- Advanced section -->
